requirements.txt
albumentations==1.3.0
click==8.0.3
google_api_python_client==2.72.0
hydra-core==1.3.1
Markdown==3.4.1
omegaconf==2.3.0
protobuf==4.21.12
python-dotenv==0.21.0
setuptools==59.6.0
timm==0.6.12
torch==1.13.1
torchvision==0.14.1
opencv-python==4.7.0.68
flask==2.1.0
pytest
coverage
wandb==0.13.9
split-folders==0.5.1
